Update README.md
Our final project for Grand Circus was an AngularJS app that utilized either an API (but NOT google maps) or a database to gather or store information. We worked in groups of four and had approximately two weeks to fulfill all of the requirements (see below). My group created a book exchange app using the Google Books and Mailjet APIs in addition to a database to update, store and match user information.